Fix renaming
